@@ -4,25 +4,25 @@
   "description": "Zentlix Main bundle",
   "license": "proprietary",
   "require": {
-    "php": ">=7.4.0",
-    "symfony/framework-bundle": "4.4.*",
-    "sensio/framework-extra-bundle": "^5.5",
-    "symfony/orm-pack": "^1.0",
-    "symfony/translation": "4.4.*",
-    "symfony/form": "4.4.*",
-    "symfony/validator": "4.4.*",
+    "php": ">=8.0.2",
+    "symfony/framework-bundle": "^5.4 || ^6.0",
+    "symfony/orm-pack": "^2.1.0",
+    "symfony/translation": "^5.4 || ^6.0",
+    "symfony/form": "^5.4 || ^6.0",
+    "symfony/validator": "^5.4 || ^6.0",
     "symfony/twig-pack": "^1.0",
-    "symfony/expression-language": "4.4.*",
-    "symfony/messenger": "4.4.*",
-    "symfony/http-client": "4.4.*",
-    "omines/datatables-bundle": "^0.4.2",
-    "tightenco/collect": "^7.9",
-    "gedmo/doctrine-extensions": "^2.4",
-    "ramsey/uuid": "^4.0",
-    "ramsey/uuid-doctrine": "^1.6",
-    "knplabs/knp-paginator-bundle": "^5.3",
-    "composer/composer": "^2.0",
-    "zentlix/user-bundle": "^1.1"
+    "symfony/expression-language": "^5.4 || ^6.0",
+    "symfony/messenger": "^5.4 || ^6.0",
+    "symfony/http-client": "^5.4 || ^6.0",
+    "symfony/uid": "^5.4 || ^6.0",
+    "symfony/monolog-bundle": "^3.7",
+    "symfony/serializer": "^5.4 || ^6.0",
+    "omines/datatables-bundle": "dev-master#e12c6756b2928f59dca2647390059dbeb329a90c",
+    "illuminate/collections": "^8.70",
+    "gedmo/doctrine-extensions": "^3.2",
+    "knplabs/knp-paginator-bundle": "dev-master#216b9d5708001788321916c5b7632da9fb9ef6ca",
+    "composer/composer": "^2.1",
+    "zentlix/user-bundle": "2.0.x-dev"
   },
   "require-dev": {
   },
@@ -42,10 +42,6 @@
   "scripts": {
     "test" : "./vendor/bin/simple-phpunit"
   },
-  "extra": {
-    "symfony": {
-      "allow-contrib": false,
-      "require": "4.4.*"
-    }
-  }
+  "minimum-stability": "dev",
+  "prefer-stable": true
 }
