@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Application\Command\Setting;
@@ -21,18 +13,15 @@ use Zentlix\MainBundle\Infrastructure\Share\Bus\DynamicPropertyCommand;
 
 class SettingCommand extends DynamicPropertyCommand implements CommandInterface, SettingCommandInterface
 {
-    /**
-     * @var string|Locale
-     * @Constraints\NotBlank()
-     */
-    public $default_locale;
+    /** @Constraints\NotBlank() */
+    public string|Locale $default_locale;
 
     private Setting $setting;
 
     public function __construct(Setting $setting)
     {
         $this->setting        = $setting;
-        $this->default_locale = $setting->getDefaultLocale()->getId()->toString();
+        $this->default_locale = $setting->getDefaultLocale()->getId()->toRfc4122();
     }
 
     public function getSettings(): SettingInterface
