@@ -1,59 +1,64 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\UI\Http\Web\Controller\Admin;
 
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController as BaseController;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\DependencyInjection\ParameterBag\ParameterBag;
+use Symfony\Component\Form\FormFactoryInterface;
+use Symfony\Component\HttpFoundation;
+use Symfony\Component\Routing\RouterInterface;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 use Throwable;
+use Twig\Environment;
 use Zentlix\MainBundle\Domain\Bundle\Service\Bundles;
-use Zentlix\MainBundle\Infrastructure\Share\Bus\CommandBus;
-use Zentlix\MainBundle\Infrastructure\Share\Bus\QueryBus;
-use Zentlix\MainBundle\Infrastructure\Share\Bus\CommandInterface;
-use Zentlix\MainBundle\Infrastructure\Share\Bus\QueryInterface;
-use Zentlix\MainBundle\UI\Http\Web\Controller\AbstractControllerInterface;
+use Zentlix\MainBundle\Infrastructure\Share\Bus;
+use Zentlix\MainBundle\UI\Http\Web\Controller\AbstractController;
 
-class AbstractAdminController extends BaseController implements AbstractControllerInterface
+abstract class AbstractAdminController extends AbstractController
 {
     public static $redirectErrorPath = 'admin.index';
 
-    private Bundles $bundles;
-    private CommandBus $commandBus;
-    private QueryBus $queryBus;
-
-    public function __construct(Bundles $bundles,
-                                CommandBus $commandBus,
-                                QueryBus $queryBus)
-    {
-        $this->bundles = $bundles;
-        $this->commandBus = $commandBus;
-        $this->queryBus = $queryBus;
+    public function __construct(
+        protected HttpFoundation\RequestStack $requestStack,
+        protected SerializerInterface $serializer,
+        protected AuthorizationCheckerInterface $authorizationChecker,
+        protected FormFactoryInterface $formFactory,
+        protected ParameterBag $parameterBag,
+        protected RouterInterface $router,
+        protected TranslatorInterface $translator,
+        protected Environment $twig,
+        protected TokenStorageInterface $tokenStorage,
+        private Bus\CommandBus $commandBus,
+        private Bus\QueryBus $queryBus,
+        private Bundles $bundles
+    ) {
+        parent::__construct(
+            $this->requestStack,
+            $this->serializer,
+            $this->authorizationChecker,
+            $this->formFactory,
+            $this->parameterBag,
+            $this->router,
+            $this->translator,
+            $this->twig,
+            $this->tokenStorage
+        );
     }
 
-    /**
-     * @throws Throwable
-     */
-    protected function exec(CommandInterface $command): void
+    /** @throws Throwable */
+    protected function exec(Bus\CommandInterface $command): void
     {
         $this->checkAccess($command);
 
         $this->commandBus->handle($command);
     }
 
-    /**
-     * @throws Throwable
-     */
-    protected function ask(QueryInterface $query)
+    /** @throws Throwable */
+    protected function ask(Bus\QueryInterface $query): mixed
     {
         $this->checkAccess($query);
 
@@ -65,7 +70,7 @@ class AbstractAdminController extends BaseController implements AbstractControll
         return ['redirect' => $url, 'success' => true, 'message' => $message];
     }
 
-    protected function redirectError(string $message): Response
+    protected function redirectError(string $message): HttpFoundation\Response
     {
         $this->addFlash('error', $message);
 
@@ -79,20 +84,20 @@ class AbstractAdminController extends BaseController implements AbstractControll
         return $this->redirectToRoute($route, $parameters);
     }
 
-    protected function jsonError(string $message): JsonResponse
+    protected function jsonError(string $message): HttpFoundation\JsonResponse
     {
         return $this->json(['success' => false, 'message' => $message]);
     }
 
-    private function checkAccess($command)
+    private function checkAccess(Bus\CommandInterface|Bus\QueryInterface $command)
     {
-        if(!$this->bundles->isBundle(get_class($command))) {
+        if(!$this->bundles->isBundle($command::class)) {
             return;
         }
 
-        $rights = $this->bundles->getBundleByNamespace(get_class($command))->configureRights();
+        $rights = $this->bundles->getBundleByNamespace($command::class)->configureRights();
 
-        if(isset($rights[get_class($command)]) && !$this->getUser()->isAccessGranted(get_class($command))) {
+        if(isset($rights[$command::class]) && !$this->getUser()->isAccessGranted($command::class)) {
             throw new \Exception('zentlix_main.access_denied');
         }
     }
