@@ -1,26 +1,18 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Application\Command\Attribute;
 
+use Symfony\Component\Uid\Uuid;
 use Zentlix\MainBundle\Infrastructure\Share\Bus\CreateCommandInterface;
-use Zentlix\MainBundle\Infrastructure\Share\Doctrine\Uuid;
 use Zentlix\MainBundle\MainBundle;
 
 class CreateCommand extends Command implements CreateCommandInterface
 {
     public function __construct(string $entity)
     {
-        $this->id              = Uuid::uuid4();
+        $this->id              = Uuid::v4();
         $this->attributeEntity = $entity;
         $this->bundle          = MainBundle::class;
     }
