@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Application\Command\Attribute;
@@ -21,14 +13,10 @@ use Zentlix\MainBundle\Infrastructure\Share\Bus\CommandHandlerInterface;
 
 class DeleteHandler implements CommandHandlerInterface
 {
-    private EntityManagerInterface $entityManager;
-    private EventDispatcherInterface $eventDispatcher;
-
-    public function __construct(EntityManagerInterface $entityManager,
-                                EventDispatcherInterface $eventDispatcher)
-    {
-        $this->entityManager = $entityManager;
-        $this->eventDispatcher = $eventDispatcher;
+    public function __construct(
+        private EntityManagerInterface $entityManager,
+        private EventDispatcherInterface $eventDispatcher
+    ) {
     }
 
     public function __invoke(DeleteCommand $command): void
