@@ -1,19 +1,11 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Application\Command\Bundle\Zentlix;
 
+use Symfony\Component\Uid\Uuid;
 use Zentlix\MainBundle\Infrastructure\Share\Bus\CommandInterface;
-use Zentlix\MainBundle\Infrastructure\Share\Doctrine\Uuid;
 use Zentlix\MainBundle\ZentlixBundleInterface;
 
 class InstallCommand implements CommandInterface
@@ -25,7 +17,7 @@ class InstallCommand implements CommandInterface
 
     public function __construct(ZentlixBundleInterface $bundle)
     {
-        $this->id           = Uuid::uuid4();
+        $this->id           = Uuid::v4();
         $this->updated_at   = new \DateTimeImmutable();
         $this->installed_at = new \DateTimeImmutable();
         $this->bundle       = $bundle;
