@@ -1,38 +1,24 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Application\Command\File;
 
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\Uid\Uuid;
 use Zentlix\MainBundle\Domain\File\Event\BeforeUpload;
 use Zentlix\MainBundle\Domain\File\Event\AfterUpload;
 use Zentlix\MainBundle\Domain\File\Service\FileUploaderInterface;
 use Zentlix\MainBundle\Domain\File\Specification\UniquePathSpecification;
 use Zentlix\MainBundle\Infrastructure\Share\Bus\CommandHandlerInterface;
-use Zentlix\MainBundle\Infrastructure\Share\Doctrine\Uuid;
 
 class UploadHandler implements CommandHandlerInterface
 {
-    private EventDispatcherInterface $eventDispatcher;
-    private FileUploaderInterface $fileUploader;
-    private UniquePathSpecification $uniquePathSpecification;
-
-    public function __construct(EventDispatcherInterface $eventDispatcher,
-                                FileUploaderInterface $fileUploader,
-                                UniquePathSpecification $uniquePathSpecification)
-    {
-        $this->eventDispatcher = $eventDispatcher;
-        $this->fileUploader = $fileUploader;
-        $this->uniquePathSpecification = $uniquePathSpecification;
+    public function __construct(
+        private EventDispatcherInterface $eventDispatcher,
+        private FileUploaderInterface $fileUploader,
+        private UniquePathSpecification $uniquePathSpecification
+    ) {
     }
 
     public function __invoke(UploadCommand $command): void
@@ -43,7 +29,7 @@ class UploadHandler implements CommandHandlerInterface
         }
 
         if($this->uniquePathSpecification->isUnique($path . '/' . $command->filename) === false) {
-            $command->filename = Uuid::uuid4() . '.' . $command->uploadedFile->getClientOriginalExtension();
+            $command->filename = Uuid::v4() . '.' . $command->uploadedFile->getClientOriginalExtension();
         }
 
         $this->eventDispatcher->dispatch(new BeforeUpload($command));
