@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Domain\DataTable\Service;
@@ -17,6 +9,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Omines\DataTablesBundle\DependencyInjection\Instantiator;
 use Omines\DataTablesBundle\Column\AbstractColumn;
 use Omines\DataTablesBundle\DataTable as BaseDataTable;
+use Omines\DataTablesBundle\Exporter\DataTableExporterManager;
 use Zentlix\MainBundle\Domain\DataTable\Entity\DataTable as DataTableConfig;
 
 class DataTable extends BaseDataTable
@@ -24,8 +17,12 @@ class DataTable extends BaseDataTable
     protected $method = Request::METHOD_GET;
     protected DataTableConfig $config;
 
-    public function __construct(EventDispatcherInterface $eventDispatcher, array $options, Instantiator $instantiator)
-    {
+    public function __construct(
+        EventDispatcherInterface $eventDispatcher,
+        array $options,
+        Instantiator $instantiator,
+        DataTableExporterManager $exporterManager
+    ) {
         // override options
         $options['searching'] = true;
         $options['pagingType'] = 'simple_numbers';
@@ -33,7 +30,7 @@ class DataTable extends BaseDataTable
         $options['pageLength'] = 25;
         $options['lengthMenu'] = [[10, 25, 50, 100, 200], [10, 25, 50, 100, 200]];
 
-        parent::__construct($eventDispatcher, $options, $instantiator);
+        parent::__construct($eventDispatcher, $exporterManager, $options, $instantiator);
     }
 
     public function setDatabaseConfig(DataTableConfig $config)
