@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Domain\Locale\Repository;
@@ -51,7 +43,7 @@ class LocaleRepository extends ServiceEntityRepository
 
         $result = [];
         foreach ($locales as $locale) {
-            $result[$locale['id']->toString()] = $locale['title'];
+            $result[$locale['id']->toRfc4122()] = $locale['title'];
         }
 
         return $result;
