@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Domain\Site\Repository;
@@ -75,7 +67,7 @@ class SiteRepository extends ServiceEntityRepository
 
         $result = [];
         foreach ($sites as $site) {
-            $result[$site['id']->toString()] = $site['title'];
+            $result[$site['id']->toRfc4122()] = $site['title'];
         }
 
         return $result;
