@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Domain\Site\Service;
@@ -16,20 +8,13 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Zentlix\MainBundle\Domain\Site\Entity\Site;
 use Zentlix\MainBundle\Domain\Cache\Service\Cache;
-use function is_null;
 
 class Sites
 {
-    private ?EntityManagerInterface $entityManager;
-    private ?string $host = null;
-
-    public function __construct(EntityManagerInterface $entityManager, RequestStack $request)
-    {
-        $this->entityManager = $entityManager;
-
-        if($request->getCurrentRequest()) {
-            $this->host = $request->getCurrentRequest()->server->get('SERVER_NAME');
-        }
+    public function __construct(
+        private EntityManagerInterface $entityManager,
+        private RequestStack $request
+    ) {
     }
 
     public function getSites()
@@ -53,29 +38,42 @@ class Sites
 
     public function getCurrentSite(): Site
     {
-        if(is_null($this->host)) {
+        $host = $this->getHost();
+
+        if(\is_null($host)) {
             throw new \DomainException('Host is null.');
         }
 
-        if(!isset($this->getSites()[$this->host])) {
+        if(!isset($this->getSites()[$host])) {
             throw new \DomainException('Site for current URL not found.');
         }
 
-        return $this->getSites()[$this->host];
+        return $this->getSites()[$host];
     }
 
     public function getCurrentSiteId(): string
     {
-        return $this->getCurrentSite()->getId()->toString();
+        return $this->getCurrentSite()->getId()->toRfc4122();
     }
 
     public function hasCurrentSite(): bool
     {
-        return is_null($this->host) === false && isset($this->getSites()[$this->host]);
+        $host = $this->getHost();
+
+        return \is_null($host) === false && isset($this->getSites()[$host]);
     }
 
     public static function clearCache(): void
     {
         Cache::clear(Cache::SITES);
+    }
+
+    private function getHost(): ?string
+    {
+        if($this->request->getCurrentRequest()) {
+            return $this->request->getCurrentRequest()->server->get('SERVER_NAME');
+        }
+
+        return null;
     }
 }
