@@ -1,17 +1,9 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Infrastructure\Collection;
 
-use Tightenco\Collect\Support\Collection as BaseCollection;
+use Illuminate\Support\Collection as BaseCollection;
 
 class Collection extends BaseCollection {}
