@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\Infrastructure\Share\Bus;
@@ -31,7 +23,7 @@ final class CommandBus
     /**
      * @throws Throwable
      */
-    public function handle($command): void
+    public function handle(CommandInterface $command): void
     {
         try {
             $this->messageBus->dispatch($command);
