@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\UI\Cli\Command;
@@ -22,13 +14,10 @@ use Zentlix\MainBundle\Infrastructure\Share\Bus\CommandBus;
 
 class CreateTemplateCommand extends ConsoleCommand
 {
-    private CommandBus $commandBus;
-
-    public function __construct(CommandBus $commandBus)
-    {
+    public function __construct(
+        private CommandBus $commandBus
+    ) {
         parent::__construct();
-
-        $this->commandBus = $commandBus;
     }
 
     protected function configure(): void
@@ -54,7 +43,7 @@ class CreateTemplateCommand extends ConsoleCommand
 
             return $folder;
         });
-        $command->sort = (int) $io->ask('Template sort', 1, function ($sort) {
+        $command->sort = (int) $io->ask('Template sort', '1', function ($sort) {
             if ((int) $sort < 1) {
                 throw new \RuntimeException('Template sort must be a positive int.');
             }
@@ -77,6 +66,6 @@ class CreateTemplateCommand extends ConsoleCommand
             "Sort: $command->sort"
         ]);
 
-        return 0;
+        return self::SUCCESS;
     }
 }
