@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\UI\Cli\Command;
@@ -20,13 +12,10 @@ use Zentlix\MainBundle\Domain\File\Service\Files;
 
 class DeleteFilesCommand extends Command
 {
-    private Files $files;
-
-    public function __construct(Files $files)
-    {
+    public function __construct(
+        private Files $files
+    ) {
         parent::__construct();
-
-        $this->files = $files;
     }
 
     protected function configure(): void
@@ -44,6 +33,6 @@ class DeleteFilesCommand extends Command
 
         $io->success('Unused files successfully deleted.');
 
-        return 0;
+        return self::SUCCESS;
     }
 }
