@@ -1,13 +1,5 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\UI\Http\Web\Controller;
@@ -17,7 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Zentlix\MainBundle\Application\Command\File\UploadCommand;
 use Zentlix\MainBundle\Application\Query\File\GetFileByIdQuery;
 
-class FileController extends AbstractController
+class FileController extends AbstractSiteController
 {
     public function upload(Request $request): Response
     {
