@@ -1,20 +1,12 @@
 <?php
 
-/**
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade Zentlix to newer
- * versions in the future. If you wish to customize Zentlix for your
- * needs please refer to https://docs.zentlix.io for more information.
- */
-
 declare(strict_types=1);
 
 namespace Zentlix\MainBundle\UI\Http\Web\Controller;
 
 use Symfony\Component\HttpFoundation\Response;
 
-class IndexController extends AbstractController
+class IndexController extends AbstractSiteController
 {
     public function index(): Response
     {
